+++ src/core/mod.rs
@@ -1,6 +1,9 @@
 #[cfg(feature = "generate")]
 pub mod generate;
 
+#[cfg(feature = "barcode")]
+pub mod barcode;
+
 #[cfg(feature = "decode")]
 pub mod decode;